+++ UI/Head.ts
@@ -1,17 +1,12 @@
 /**
  * Библиотека заголовков страницы
  * @library UI/Head
+ * @includes applyHeadJson UI/_head/defaultTags
+ * @includes createMetaScriptsAndLinks UI/_head/defaultTags
  * @public
  * @author Печеркин С.В.
  */
-import { createWsConfig } from 'UI/_head/wsConfig';
-import { createTitle, createViewPort, createDefaultTags, createMetaScriptsAndLinks, applyHeadJson, IHeadOptions } from 'UI/_head/defaultTags';
-export {
-   applyHeadJson,
-   createTitle,
-   createViewPort,
-   createWsConfig,
-   createDefaultTags,
-   createMetaScriptsAndLinks,
-   IHeadOptions
-};
+export { createWsConfig } from 'UI/_head/wsConfig';
+export { createTitle, createViewPort, createDefaultTags,
+         createMetaScriptsAndLinks, applyHeadJson } from 'UI/_head/defaultTags';
+export { IHeadOptions } from 'UI/_head/Interface';
