+++ UI/_head/Interface.ts
@@ -1,24 +1,23 @@
 /**
  * Интерфейс для контролов, поддерживающих конфигурацию <a href="/doc/platform/developmentapl/middleware/preprocessor/">сервиса представления</a> приложения.
- * @interface UI/_base/interface/IRootTemplate
- * @public
+ * @interface UI/_head/Interface:IRootTemplateOptions
  * @author Санников К.А.
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#buildnumber
+ * @name UI/_head/Interface:IRootTemplateOptions#buildnumber
  * @cfg {string} Номер сборки. Прибавляется ко всем url при запросах к серверу.
  * Используется при кешировании. Если меняется номер сборки,
  * все файлы начинают тянуться с сервера, а не из кеша браузера, потому что у них становится другой url.
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#product
+ * @name UI/_head/Interface:IRootTemplateOptions#product
  * @cfg {string}
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#wsRoot
+ * @name UI/_head/Interface:IRootTemplateOptions#wsRoot
  * @cfg {String} Путь к корню интерфейсного модуля WS.Core. Например, "/resources/WS.Core/".
  * @remark
  * Значение опции задаётся относительно URL-адреса сервиса.
@@ -27,12 +26,7 @@
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#wsRoot
- * @cfg {String} Path to ws root url
- */
-
-/**
- * @name UI/_base/interface/IRootTemplate#resourceRoot
+ * @name UI/_head/Interface:IRootTemplateOptions#resourceRoot
  * @cfg {String} Адрес к директории с ресурсами сервиса. Например, "/resources/".
  * @remark
  * Значение опции задаётся относительно URL-адреса сервиса.
@@ -41,12 +35,7 @@
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#resourceRoot
- * @cfg {String} Path to resource root url
- */
-
-/**
- * @name UI/_base/interface/IRootTemplate#appRoot
+ * @name UI/_head/Interface:IRootTemplateOptions#appRoot
  * @cfg {String} Адрес к директории сервиса. Например, "/".
  * @remark
  * Значение опции задаётся относительно URL-адреса сервиса.
@@ -55,29 +44,18 @@
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#appRoot
- * @cfg {String} Path to application root url
- */
-
-/**
- * @name UI/_base/interface/IRootTemplate#staticDomains
+ * @name UI/_head/Interface:IRootTemplateOptions#staticDomains
  * @cfg {Array} Список, содержащий набор доменов для загрузки статики.
  * Список доменов решает задачу загрузки статических ресурсов с нескольких документов. Эти домены будут использоваться для создания путей для статических ресурсов и распределения загрузки для нескольких статических доменов.
  */
 
 /**
- * @name UI/_base/interface/IRootTemplate#staticDomains
- * @cfg {Array} The list of domains for distributing static resources. These domains will be used to create paths
- * for static resources and distribute downloading for several static domains.
- */
-
-/**
- * @name UI/_base/interface/IRootTemplate#servicesPath
+ * @name UI/_head/Interface:IRootTemplateOptions#servicesPath
  * @cfg {string} Имя по-умолчанию для обращения к сервисам бизнес-логики
  * Данная настройка попадает в свойство defaultServiceUrl объекта window.wsConfig.
  */
 
-export interface IRootTemplateOptions {
+interface IRootTemplateOptions {
     buildnumber?: string;
     product?: string;
     wsRoot?: string;
@@ -85,4 +63,22 @@ export interface IRootTemplateOptions {
     appRoot?: string;
     staticDomains?: string[];
     servicesPath?: string;
+    pageName?: string;
+    RUMEnabled?: boolean;
+}
+
+/**
+ * Интерфейс описывающий опции, которые необходимы при формировании содержимого head и скриптов страницы
+ * @interface UI/_head/Interface:IHeadOptions
+ * @public
+ */
+export interface IHeadOptions extends IRootTemplateOptions {
+    defaultTheme?: string;
+    theme?: string;
+    noscript?: string;
+    preInitScript?: string;
+    reactApp?: boolean;
+    meta?: Object[];
+    links?: Object[];
+    scripts?: Object[];
 }
