+++ UI/_head/defaultTags.ts
@@ -4,21 +4,8 @@ import { Head as AppHead, JSLinks } from 'Application/Page';
 import type { JML } from 'Application/Page';
 import { getResourceUrl } from 'UI/Utils';
 import escapeHtml = require('Core/helpers/String/escapeHtml');
+import { IHeadOptions } from 'UI/_head/Interface';
 
-import { IRootTemplateOptions } from '../_base/interface/IRootTemplate';
-
-export interface IHeadOptions extends IRootTemplateOptions {
-   defaultTheme?: string;
-   theme?: string;
-   noscript?: string;
-   preInitScript?: string;
-   reactApp?: boolean;
-   pageName?: string;
-   RUMEnabled?: boolean;
-   meta?: Object[];
-   links?: Object[];
-   scripts?: Object[];
-}
 
 export function createTitle(title: string): void {
    AppHead.getInstance().createTag('title', {}, title);
